fixup! Método para consulta de NFes movido para res.company
Método "_mask_cnpj" movido para nova classe correta
