+++ nfe_mde/nfe_schedule.py
@@ -17,8 +17,6 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.       #
 ###############################################################################
 
-import re
-
 from openerp import models, api, fields
 
 
@@ -30,15 +28,6 @@ class NfeSchedule(models.TransientModel):
         selection=[('init', 'Não iniciado'), ('done', 'Finalizado')],
         default='init'
     )
-
-    @staticmethod
-    def _mask_cnpj(cnpj):
-        if cnpj:
-            val = re.sub('[^0-9]', '', cnpj)
-            if len(val) == 14:
-                cnpj = "%s.%s.%s/%s-%s" % (val[0:2], val[2:5], val[5:8],
-                                           val[8:12], val[12:14])
-        return cnpj
 
     @api.model
     def schedule_download(self, raise_error=False, domain=()):
